refactor (#60) : 패키지 경로 이동
